added the ability to edit and delete data

+++ src/main/java/com/example/cometa/domain/Color.java
@@ -56,7 +56,7 @@ public class Color {
         return value;
     }
 
-    public void setValue(float value) {
+    public void setValue(double value) {
         this.value = value;
     }
 }
